+++ business-logic/src/main/java/com/netflix/vms/transformer/hollowoutput/RolloutVideo.java
@@ -49,5 +49,5 @@ public class RolloutVideo implements Cloneable {
     }
 
     @SuppressWarnings("unused")
-    private int __assigned_ordinal = -1;
+    private long __assigned_ordinal = -1;
 }
